notebook + output for Comedy Central web scraping
Here is the notebook for the web scraping for Comedy Central, however I also did all the scraping (it took a LONG time) and saved all the jokes to a csv. It seems there are some repeats (same joke will appear in multiple categories, with a new URL each time, although the last part of the URL path remains the same--it's the joke title, apparently). My notebook won't run all the way through as is -- I had to run the scraping in stages, and just modified some values in the functions as I did that. That's why I saved out outputs for you!

Two csv's are (1) the jokes output, plus (2) the list of all URLs I scraped from the topic pages.

I think some jokes/URLs got passed over in the scraping, but I can't really tell easily because of the joke multi-categorization. You might be able to figure it out more easily once bring them back into a dataframe.